split core conversion functions into separate file
