+++ kimchi/js/bindings.js
@@ -946,7 +946,7 @@ var caml_pasta_fp_plonk_gate_vector_add = function (v, x) {
 // Provides: caml_pasta_fp_plonk_gate_vector_get
 // Requires: plonk_wasm, tsRustConversion
 var caml_pasta_fp_plonk_gate_vector_get = function (v, i) {
-    return tsRustConversion.gateFromRust(plonk_wasm.caml_pasta_fp_plonk_gate_vector_get(v, i));
+    return tsRustConversion.fp.gateFromRust(plonk_wasm.caml_pasta_fp_plonk_gate_vector_get(v, i));
 };
 
 // Provides: caml_pasta_fp_plonk_gate_vector_len
@@ -993,7 +993,7 @@ var caml_pasta_fq_plonk_gate_vector_add = function (v, x) {
 // Provides: caml_pasta_fq_plonk_gate_vector_get
 // Requires: plonk_wasm, tsRustConversion
 var caml_pasta_fq_plonk_gate_vector_get = function (v, i) {
-    return tsRustConversion.gateFromRust(plonk_wasm.caml_pasta_fq_plonk_gate_vector_get(v, i));
+    return tsRustConversion.fq.gateFromRust(plonk_wasm.caml_pasta_fq_plonk_gate_vector_get(v, i));
 };
 
 // Provides: caml_pasta_fq_plonk_gate_vector_len
